pset 4/requirements.txt
Pillow==8.3.2
matplotlib==3.3.3
mercantile==1.2.1
numpy==1.20.1
requests==2.26.0
dbf==0.99.1
dbfread==2.0.7
pyproj==3.2.1
